+++ CHANGELOG.md
@@ -70,6 +70,8 @@ more scanner types.
 
 "MQ is for Message Queue"
 
+- Fixed an issue where an LDAP import job would show as successful despite finding no users
+
 ### New in this release
 
 - Organizations can now define a retention policy. This policy replaces the previously hardcoded